Add Judit result parsing to process detail view
